# kita-ax/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KYRA Admin Console API",
    "description": "REST API for KYRA AI Guardrail Admin Console with comprehensive user, document, policy, and audit management",
    "version": "1.0.0",
    "contact": {
      "name": "SeekersLab",
      "url": "https://seekerslab.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://127.0.0.1:3005/api",
      "description": "Development server",
      "variables": {
        "version": {
          "default": "v1",
          "enum": ["v1", "v2"]
        }
      }
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "user-001" },
          "email": { "type": "string", "format": "email", "example": "admin@seekerslab.com" },
          "role": { "type": "string", "enum": ["admin", "editor", "viewer"], "example": "admin" },
          "tenantId": { "type": "string", "example": "tenant-001" },
          "status": { "type": "string", "enum": ["active", "inactive"], "example": "active" },
          "lastLogin": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "email", "role", "status", "createdAt"]
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "newuser@seekerslab.com" },
          "role": { "type": "string", "enum": ["admin", "editor", "viewer"], "example": "editor" }
        },
        "required": ["email", "role"]
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "editor", "viewer"] },
          "status": { "type": "string", "enum": ["active", "inactive"] }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "doc-001" },
          "title": { "type": "string", "example": "Security Policy v2.0" },
          "classification": { "type": "string", "enum": ["public", "internal", "confidential", "secret"] },
          "owner": { "type": "string", "format": "email" },
          "uploadedAt": { "type": "string", "format": "date-time" },
          "accessCount": { "type": "integer", "minimum": 0 },
          "size": { "type": "integer", "minimum": 0 }
        },
        "required": ["id", "title", "classification", "owner", "uploadedAt"]
      },
      "DocumentCreate": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "classification": { "type": "string", "enum": ["public", "internal", "confidential", "secret"] },
          "owner": { "type": "string", "format": "email" }
        },
        "required": ["title", "classification", "owner"]
      },
      "Role": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "admin" },
          "description": { "type": "string" },
          "permissions": { "type": "array", "items": { "type": "string" } },
          "userCount": { "type": "integer", "minimum": 0 }
        },
        "required": ["name", "description", "permissions"]
      },
      "ABACRule": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Executive Access" },
          "condition": { "type": "string", "example": "department == 'executive'" },
          "effect": { "type": "string", "enum": ["allow", "deny"] },
          "resources": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "enum": ["active", "inactive"] }
        },
        "required": ["name", "condition", "effect", "resources"]
      },
      "Policy": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["rbac", "abac"] },
          "target": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["name", "type", "target", "status", "createdAt"]
      },
      "AuditLog": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "eventType": { "type": "string", "enum": ["authentication", "document-access", "policy-change", "user-management"] },
          "user": { "type": "string", "format": "email" },
          "resource": { "type": "string" },
          "action": { "type": "string" },
          "status": { "type": "string", "enum": ["success", "failure"] },
          "ipAddress": { "type": "string", "format": "ipv4" }
        },
        "required": ["timestamp", "eventType", "user", "resource", "action", "status"]
      },
      "Agent": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["analysis", "automation"] },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "apiKey": { "type": "string" },
          "lastSeen": { "type": "string", "format": "date-time" },
          "requests24h": { "type": "integer", "minimum": 0 }
        },
        "required": ["name", "type", "status", "apiKey", "lastSeen"]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "minimum": 1 },
          "pageSize": { "type": "integer", "minimum": 1, "maximum": 100 },
          "total": { "type": "integer", "minimum": 0 },
          "totalPages": { "type": "integer", "minimum": 0 },
          "hasNext": { "type": "boolean" },
          "hasPrev": { "type": "boolean" }
        },
        "required": ["page", "pageSize", "total", "totalPages", "hasNext", "hasPrev"]
      },
      "PaginatedUsers": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        },
        "required": ["data", "pagination"]
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "type": "object" },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["success", "data", "timestamp"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "code": { "type": "string" },
              "timestamp": { "type": "string", "format": "date-time" }
            }
          }
        },
        "required": ["success", "error"]
      }
    },
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId",
        "description": "Session cookie authentication"
      }
    }
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "List users",
        "description": "Get paginated list of users with filtering, searching, and sorting",
        "tags": ["Users"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search by email or role" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "inactive"] } },
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "editor", "viewer"] } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "asc" } }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedUsers" } } }
          },
          "401": { "description": "Unauthorized" },
          "400": { "description": "Bad request" }
        }
      },
      "post": {
        "summary": "Create user",
        "description": "Create a new user account",
        "tags": ["Users"],
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } } }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user",
        "description": "Get a specific user by ID",
        "tags": ["Users"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "User found" },
          "404": { "description": "User not found" },
          "401": { "description": "Unauthorized" }
        }
      },
      "put": {
        "summary": "Update user",
        "description": "Update user information",
        "tags": ["Users"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } } }
        },
        "responses": {
          "200": { "description": "User updated" },
          "400": { "description": "Validation error" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "summary": "Delete user",
        "description": "Delete a user account",
        "tags": ["Users"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/documents": {
      "get": {
        "summary": "List documents",
        "description": "Get paginated list of documents",
        "tags": ["Documents"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "classification", "in": "query", "schema": { "type": "string", "enum": ["public", "internal", "confidential", "secret"] } }
        ],
        "responses": {
          "200": { "description": "List of documents" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create document",
        "tags": ["Documents"],
        "security": [{ "sessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentCreate" } } }
        },
        "responses": {
          "201": { "description": "Document created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/documents/{id}": {
      "get": {
        "summary": "Get document",
        "tags": ["Documents"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Document found" },
          "404": { "description": "Document not found" }
        }
      },
      "put": {
        "summary": "Update document",
        "tags": ["Documents"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Document updated" },
          "404": { "description": "Document not found" }
        }
      },
      "delete": {
        "summary": "Delete document",
        "tags": ["Documents"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Document deleted" },
          "404": { "description": "Document not found" }
        }
      }
    },
    "/roles": {
      "get": {
        "summary": "List roles",
        "tags": ["Roles"],
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": { "description": "List of roles" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create role",
        "tags": ["Roles"],
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "201": { "description": "Role created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/roles/{name}": {
      "get": {
        "summary": "Get role",
        "tags": ["Roles"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Role found" },
          "404": { "description": "Role not found" }
        }
      }
    },
    "/audit-logs": {
      "get": {
        "summary": "List audit logs",
        "description": "Get audit logs with filtering and search",
        "tags": ["Audit"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "eventType", "in": "query", "schema": { "type": "string", "enum": ["authentication", "document-access", "policy-change", "user-management"] } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["success", "failure"] } }
        ],
        "responses": {
          "200": { "description": "List of audit logs" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/dashboard/metrics": {
      "get": {
        "summary": "Get dashboard metrics",
        "description": "Get KPI metrics for dashboard",
        "tags": ["Dashboard"],
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": {
            "description": "Dashboard metrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalDocuments": { "type": "integer" },
                        "activeUsers": { "type": "integer" },
                        "totalPolicies": { "type": "integer" },
                        "auditEvents": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "tags": [
    { "name": "Users", "description": "User management endpoints" },
    { "name": "Documents", "description": "Document management endpoints" },
    { "name": "Roles", "description": "Role and RBAC endpoints" },
    { "name": "Audit", "description": "Audit log endpoints" },
    { "name": "Dashboard", "description": "Dashboard metrics endpoints" }
  ]
}